# swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "FOODIE-FUND",
    "description": "眾資成城-在你心愛的餐廳成為合夥人"
  },
  "servers": [
    {
      "url": "https://foodiefund.onrender.com",
      "description": "Render"
    },
    {
      "url": "http://127.0.0.1:2330",
      "description": "Local"
    }
  ],
...
  "paths": {
...
    "/v1/api/admin/account/": {
      "get": {
        "tags": [
          "會員管理"
        ],
        "summary": "會員清單",
        "description": "會員清單",
        "parameters": [
          {
            "name": "role",
            "in": "query",
            "description": "管理身分, user,admin",
            "enum": [
              "user",
              "admin",
              "undefined"
            ],
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isBlackListed",
            "in": "query",
            "description": "1:搜尋黑名單 0:不是黑名單, 預設空直為搜尋全部",
            "enum": [
              1,
              0,
              "undefined"
            ],
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "keyWord",
            "in": "query",
            "description": "關鍵字fuzzy[email,phone,address,dateOfBirth], 預設空直為搜尋全部",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "timeSort",
            "in": "query",
            "description": "排序遠到近,進到遠default = desc",
            "enum": [
              "asc",
              "desc"
            ],
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "清單顯示比數,default=10",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "顯示第幾頁資料default=1",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/v1/api/admin/upload": {
      "post": {
        "tags": [
          "圖片上傳"
        ],
        "summary": "圖片上傳",
        "description": "圖片上傳",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "imageUrl": {
                      "type": "string",
                      "example": "image URL(https)"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "imageUrl": {
                      "type": "string",
                      "example": "image URL(https)"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Error Msg"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Error Msg"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "description": "貼文牆",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "file"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}